// src/modules/project/constants/service-type.enum.ts
export enum ServiceType {
  MarketResearch = "market_research",
  LegalServices = "legal_services",
  TaxConsulting = "tax_consulting",
  BusinessSetup = "business_setup",
  LocalPartnerships = "local_partnerships",
  RegulatoryCompliance = "regulatory_compliance",
  Translation = "translation",
  OfficeSetup = "office_setup",
}